Refactor RuleInputPanels and RuleCalculationPanels to use the same Panels wrapper
